chore: seperate layout header and the page
